events: display: make the format a struct
DisplayFormat was an enum, which does not modelize well its
capabilities. The format could embed more specific information than a
generic flavor, eg. to control the display of a given type or to give
more control to a sub-section about how the formatting should be done.

This makes it a struct to allow embedding various kinds of information.
We're still keeping the overall flavor (single & multi line) for now as
this top level configuration seems to make sense.

Signed-off-by: Antoine Tenart <atenart@redhat.com>