collect: use a default set of probes when none are explicitly given
This makes `retis collect` to use a default set of probes when none were
given explicitly. Those probes are net:netif_receive_skb and
net:net_dev_start_xmit (in addition to the ones installed by collectors,
eg. the skb-drop one) to mimic AF_PACKET utilities and because that's a
good starting point.

In addition make sure those automatic probes are only set if no profile
is used and not collector is explicitly enabled. This is important as a
profile behavior should be predictable.

Signed-off-by: Antoine Tenart <atenart@redhat.com>